moves and edits for merging
