+++ pom.xml
@@ -81,18 +81,7 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-			<groupId>org.slf4j</groupId>
-			<artifactId>slf4j-api</artifactId>
-			<version>1.7.13</version>
-		</dependency>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>4.12</version>
-			<scope>test</scope>
-   	</dependency>
-		<dependency>
-				<groupId>io.springfox</groupId>
+			<groupId>io.springfox</groupId>
 			<artifactId>springfox-swagger-ui</artifactId>
 			<version>3.0.0-SNAPSHOT</version>
 		</dependency>
@@ -141,8 +130,8 @@
           <archive>
             <manifest>
               <addClasspath>true</addClasspath>
-              <mainClass>com.microsoft.cse.helium.HeliumJavaApplication</mainClass>
-              <addDefaultImplementationEntries>true</addDefaultImplementationEntries>
+              <mainClass>com.microsoft.cse.helium.app.HeliumJavaApplication</mainClass>
+              <addDefaultImplementationEntries>false</addDefaultImplementationEntries>
             </manifest>
             <manifestEntries>
               <Implementation-Build>$\{buildNumber}</Implementation-Build>
