+++ src/main/java/com/microsoft/cse/helium/app/controllers/SecretController.java
@@ -5,7 +5,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.server.ServerHttpResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,11 +12,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import reactor.core.publisher.Mono;
 
-import com.microsoft.azure.AzureEnvironment;
-import com.microsoft.azure.credentials.MSICredentials;
-import com.microsoft.azure.credentials.AzureCliCredentials;
-import com.microsoft.azure.keyvault.KeyVaultClient;
-import com.microsoft.azure.keyvault.models.SecretBundle;
 import com.microsoft.cse.helium.app.services.keyvault.KeyVaultService;
 
 @RestController
