--- src/test/java/com/microsoft/cse/helium/app/demo/HeliumJavaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.microsoft.cse.helium.app.demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-//@SpringBootTest
-class HeliumJavaApplicationTests {
-
-	//@Test
-	void contextLoads() {
-	}
-
-}
